requirements.txt
# P-WOS Simulation Requirements
# Install with: pip install -r requirements.txt

# MQTT Communication
paho-mqtt>=1.6.1

# Web API Framework
Flask>=2.3.0
flask-cors>=4.0.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# Machine Learning
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
